Fix tests causing db access to be inconsistent
This was being caused by tests that create/delete tables interweaving
and leaving the database in an inconsistent state.

I tried using the 'runInBand' option for jest, but this didn't
consistently fix the problem.

The solution I eventually came up with (and it feels a bit hacky), is
to have each test work on a different table name so there are no
overlaps.