Continue working on addTrailersToDb
Note this build is currently broken, but I wanted to make a commit
as I'm not sure when I'm next going to get the chance to work on
this project.

The function works when used by the tests, but something is stopping
it from accessing the local instance of dynamoDB when invoked from
SAM. I'm fairly sure it's either to do with IAM permissions or the
fact that SAM runs within a Docker container and doesn't know where
then db instance is, but haven't been able to narrow it down beyond
that so far.